Add more modular http structure
