+++ backend/Cargo.toml
@@ -24,7 +24,7 @@ reqwest = "0.8.6"
 
 [dependencies]
 actix = "0.7.2"
-actix-web = { version = "0.7.1", features = ["alpn"] }
+actix-web = { version = "=0.7.1", features = ["alpn"] }
 bytes = "0.4.9"
 diesel = { version = "1.3.2", features = ["r2d2", "postgres"] }
 env_logger = "0.5.11"
